+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Sin Publicar
+- Added Multiple Row Component
 - Resized pill icon on Hybrid Carousel
 
 # v1.18.0